Revert back to List<Long> for the user's age range

+++ app/src/main/java/ch/epfl/sdp/blindly/user/User.kt
@@ -258,7 +258,7 @@ class User private constructor(
                 val matches = get("matches") as? List<String>
                 val likes = get("likes") as? List<String>
                 val description = getString("description")!!
-                val ageRange = get("ageRange") as? List<Int>
+                val ageRange = get("ageRange") as? List<Long>
                 val recordingPath = getString("recordingPath")!!
 
                 return User(
